requirements.txt
alembic==1.3.2
Click==7.0
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.12
Werkzeug==0.16.0
